Split into separate files. Rename tree-sitter to modulrity_check.py
